requirements.txt
GitPython
sparecores-data
sqlmodel
docker
sparecores-crawler